refactor: use u128 for Price and Quantity minor units

+++ tradingengine/src/lib.rs
@@ -8,6 +8,6 @@ pub mod resting_order;
 
 pub const SCALE: u32 = 8;
 
-pub const ONE: i64 = 100_000_000;
+pub const ONE: u128 = 100_000_000;
 
-const _: () = assert!(ONE == 10i64.pow(SCALE));
+const _: () = assert!(ONE == 10u128.pow(SCALE));
