+++ tradingengine/src/price.rs
@@ -1,10 +1,10 @@
 use crate::error::{EngineError, Result};
 
 #[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord, Hash)]
-pub struct Price(i64);
+pub struct Price(u128);
 
 impl Price {
-    pub fn from_minor_units(minor_units: i64) -> Result<Self> {
+    pub fn from_minor_units(minor_units: u128) -> Result<Self> {
         if minor_units > 0 {
             Ok(Price(minor_units))
         } else {
@@ -12,7 +12,7 @@ impl Price {
         }
     }
 
-    pub fn minor_units(self) -> i64 {
+    pub fn minor_units(self) -> u128 {
         self.0
     }
 }
